requirements.txt
alembic==1.16.1
apispec==6.8.2
biopython==1.86
blinker==1.9.0
click==8.2.1
Flask==3.1.2
flask-cors==6.0.0
Flask-Migrate==4.1.0
flask-smorest==0.46.2
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.2.0
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.4.0
pytest-env==1.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
webargs==8.7.0
Werkzeug==3.1.3
